django-backend old/requirements/dev.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2.0,<4.3
pytest-django>=4.7.0,<4.8
black>=23.0.0,<24.0
flake8>=6.0.0,<7.0
